+++ Sample-01/Sources/Shared/ProfileView.swift
@@ -1,0 +1,15 @@
+import Foundation
+import SwiftUI
+
+struct ProfileView: View {
+    @Binding var profile: Profile
+
+    var body: some View {
+        VStack(spacing: 16) {
+            AsyncImage(url: URL(string: profile.picture))
+            Text(profile.id)
+            Text(profile.email)
+            Text(profile.updatedAt)
+        }
+    }
+}
